pyproject.toml
[tool.poetry]
name = "hw1"
version = "0.1.0"
description = "My very first homework on MLOps course @ FTDA"
authors = ["Armine Titizian <takkat14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
pandas = "^1.5.1"
flask = "^2.2.2"
gunicorn = "^20.1.0"
gdown = "^4.5.3"
hydra-core = "^1.2.0"
scikit-learn = "^1.1.3"
flask-restx = "^1.0.3"
joblib = "^1.2.0"
pymongo = "^4.3.3"
minio = "^7.1.13"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
